+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Add `LicenceHeader` pre-commit enforcement to ensure open source projects
   contain proper licence comments.
 * Add [`Foodcritic`](http://www.foodcritic.io/) pre-commit hook
+* Fix `CapitalizedSubject` to not fail when commit message starts with one or
+  more empty lines
 
 ## 0.36.0